Move savegame serialization functions out into their own cross platform source
This file should be platform agnostic (i.e. no Windows headers!)
