+++ src/Windows Code Release 3/fileio.cpp
@@ -24,6 +24,7 @@
 #include "graphutl.h"
 #include "exlsound.h"
 #include "endian_adjust.hpp"
+#include "savedata_serialization.hpp"
 
 
 HWND	the_dialog;
@@ -108,21 +109,6 @@ static const std::array szFilter{ "Blades of Exile Save Files (*.SAV)","*.sav",
 	"All Files (*.*)","*.*",
 	"" };
 
-class boe_error
-	: public std::exception
-{
-	short error_code_;
-
-public:
-	boe_error(short error_code)
-		: std::exception(std::format("BOE ERROR #{}", error_code).c_str())
-		, error_code_(error_code)
-	{
-	}
-	short error_code() const { return error_code_; }
-};
-
-
 static bool FSClose(auto& file)
 {
 	file.close();
@@ -141,6 +127,7 @@ static bool file_read_type(auto& file, auto& type)
 	return !file.fail();
 }
 
+#if 0
 static void stream_read_type(std::istream& file, auto& type)
 {
 	assert((file.exceptions() & std::ios_base::failbit) == std::ios_base::failbit);
@@ -155,6 +142,7 @@ static T stream_read_type(std::istream& file)
 	file.read(reinterpret_cast<char*>(&type), sizeof(type));
 	return type;
 }
+#endif
 
 static void file_read_string(std::ifstream& file, long len, char* arr)
 {
@@ -168,46 +156,13 @@ static bool file_write_type(auto& file, const auto& type)
 	return !file.fail();
 }
 
+#if 0
 static void stream_write_type(auto& file, const auto& type)
 {
 	assert((file.exceptions() & std::ios_base::failbit) == std::ios_base::failbit);
 	file.write(reinterpret_cast<const char*>(&type), sizeof(type));
 }
-
-
-static void xor_type(auto& type, char xor_value)
-{
-	for (size_t index = 0; index < sizeof(type); ++index)
-	{
-		reinterpret_cast<char*>(&type)[index] ^= xor_value;
-	}
-}
-
-
-template< char K, typename T>
-struct xor_wrap_detail
-{
-	T& t_;
-	xor_wrap_detail(T & t)
-		: t_(t)
-	{
-		xor_type(t_, K);
-	}
-	T& get() const
-	{
-		return t_;
-	}
-	~xor_wrap_detail()
-	{
-		xor_type(t_, K);
-	}
-};
-
-template<char K, typename T>
-xor_wrap_detail<K, T> xor_wrap(T& t)
-{
-	return xor_wrap_detail<K, T>(t);
-}
+#endif
 
 void file_initialize()
 {
@@ -234,88 +189,6 @@ void file_initialize()
 	ofn.lCustData = 0L;
 	ofn.lpfnHook = NULL;
 	ofn.lpTemplateName = NULL;
-}
-
-static bool savedata_read_flag(std::istream& file_id, flag_type value_true, flag_type value_false)
-{
-	const auto flag = stream_read_type<flag_type>(file_id);
-	if ((flag != value_true) && (flag != value_false)) { // OK Exile II save file?
-		throw boe_error(1063);
-	}
-	return flag == value_true;
-}
-
-
-static void savedata_read_all(
-	std::istream& file_id,
-	bool & town_restore,
-	bool & in_scen,
-	party_record_type & party,
-	setup_save_type & setup_save,
-	std::array<pc_record_type, 6> & adven,
-	unsigned char (&out_e)[96][96],
-	current_town_type& c_town,
-	big_tr_type& t_d,
-	town_item_list& t_i,
-	stored_items_list_type (&stored_items)[3],
-	stored_town_maps_type& town_maps,
-	stored_town_maps_type& town_maps2,
-	stored_outdoor_maps_type& o_maps,
-	unsigned char(&sfx)[64][64],
-	unsigned char (&misc_i)[64][64]
-	)
-{
-	// Eventually this will be hidden by an interface so this kind of
-	// ugliness will go away.
-	if ((file_id.exceptions() & std::ios_base::failbit) != std::ios_base::failbit)
-	{
-		throw std::runtime_error("Internal error: called without the stream first being set to throw on failure");
-	}
-
-	town_restore = savedata_read_flag(file_id, flag_type::town, flag_type::out);
-	in_scen = savedata_read_flag(file_id, flag_type::in_scenario, flag_type::not_in_scenario);
-	const bool maps_there = savedata_read_flag(file_id, flag_type::have_maps, flag_type::no_maps);
-
-	// LOAD PARTY
-	stream_read_type(file_id, party);
-	xor_type(party, 0x5C);
-
-	// LOAD SETUP
-	stream_read_type(file_id, setup_save);
-
-	// LOAD PCS
-	stream_read_type(file_id, adven);
-	xor_type(adven, 0x6B);
-
-	if (in_scen) {
-
-		// LOAD OUTDOOR MAP
-		static_assert(sizeof(out_info_type) == sizeof(out_e));
-		stream_read_type(file_id, out_e);
-
-		// LOAD TOWN 
-		if (town_restore) {
-			stream_read_type(file_id, c_town);
-			stream_read_type(file_id, t_d);
-			stream_read_type(file_id, t_i);
-		}
-
-		// LOAD STORED ITEMS
-		stream_read_type(file_id, stored_items);
-
-		// LOAD SAVED MAPS
-		if (maps_there) {
-			stream_read_type(file_id, town_maps);
-			stream_read_type(file_id, town_maps2);
-			stream_read_type(file_id, o_maps);
-		}
-
-		// LOAD SFX & MISC_I
-		static_assert(sizeof(sfx) == 64 * 64);
-		stream_read_type(file_id, sfx);
-		static_assert(sizeof(misc_i) == 64 * 64);
-		stream_read_type(file_id, misc_i);
-	}
 }
 
 void load_file()
@@ -461,79 +334,12 @@ void load_file()
 
 }
 
-static void savedata_write_all(
-	std::ostream& file_id,
-	bool write_town_data,
-	bool write_scenario_data,
-	bool write_map_data,
-	party_record_type& party,
-	const setup_save_type& setup_save,
-	std::array<pc_record_type, 6>& adven,
-	const unsigned char(&out_e)[96][96],
-	const current_town_type& c_town,
-	const big_tr_type& t_d,
-	const town_item_list& t_i,
-	const stored_items_list_type(&stored_items)[3],
-	const stored_town_maps_type& town_maps,
-	const stored_town_maps_type& town_maps2,
-	const stored_outdoor_maps_type& o_maps,
-	const unsigned char(&sfx)[64][64],
-	const unsigned char(&misc_i)[64][64]
-	)
-{
-	// Eventually this will be hidden by an interface so this kind of
-	// ugliness will go away.
-	if ((file_id.exceptions() & std::ios_base::failbit) != std::ios_base::failbit)
-	{
-		throw std::runtime_error("Internal error: called without the stream first being set to throw on failure");
-	}
-
-	stream_write_type(file_id, write_town_data ? flag_type::town : flag_type::out);
-	stream_write_type(file_id, write_scenario_data ? flag_type::in_scenario : flag_type::not_in_scenario);
-	stream_write_type(file_id, write_map_data ? flag_type::have_maps : flag_type::no_maps);
-
-	// SAVE PARTY
-	stream_write_type(file_id, xor_wrap<0x5C>(party).get());
-
-	// SAVE SETUP
-	stream_write_type(file_id, setup_save);
-
-	// SAVE PCS	
-	stream_write_type(file_id, xor_wrap<0x6B>(adven).get());
-
-	if (write_scenario_data) {
-		// SAVE OUT DATA
-		static_assert(sizeof(out_info_type) == sizeof(out_e));
-		stream_write_type(file_id, out_e);
-
-		if (write_town_data) {
-			stream_write_type(file_id, c_town);
-			stream_write_type(file_id, t_d);
-			stream_write_type(file_id, t_i);
-		}
-
-		// Save stored items 
-		stream_write_type(file_id, stored_items);
-
-		// If saving maps, save maps
-		if (write_map_data) {
-			stream_write_type(file_id, town_maps);
-			stream_write_type(file_id, town_maps2);
-			stream_write_type(file_id, o_maps);
-		}
-
-		// SAVE SFX and MISC_I
-		static_assert(sizeof(sfx) == 64 * 64);
-		stream_write_type(file_id, sfx);
-		static_assert(sizeof(misc_i) == 64 * 64);
-		stream_write_type(file_id, misc_i);
-	}
-}
-
 void save_file(short mode)
 //mode;  // 0 - normal  1 - save as
 {
-	const bool town_save = (in_startup_mode == FALSE) && is_town();
+	Boolean town_save = FALSE;
+	if ((in_startup_mode == FALSE) && (is_town()))
+		town_save = TRUE;
 
 	ofn.hwndOwner = mainPtr;
 	ofn.lpstrFile = szFileName;
@@ -559,7 +365,7 @@ void save_file(short mode)
 	{
 		file_id.exceptions(std::ios_base::failbit);
 
-		savedata_write_all(file_id, town_save, in_startup_mode == FALSE, save_maps == TRUE,
+		savedata_write_all(file_id, town_save, in_startup_mode, save_maps,
 			party, setup_save, adven, out_e, c_town, t_d, t_i,
 			stored_items, town_maps, town_maps2, o_maps, sfx, misc_i);
 	}
